+++ src/CMakeLists.txt
@@ -17,5 +17,4 @@ INSTALL(FILES "${MY_INSTALL_SRCDIR}/configure.hpp"
 
 add_subdirectory(tbox)
 add_subdirectory(cdg)
-
 add_subdirectory(apps)
